refactor monitoring
